+++ sdk/synapse/Azure.Analytics.Synapse.Spark/src/Customization/SparkBatchOperation.cs
@@ -146,7 +146,6 @@ namespace Azure.Analytics.Synapse.Spark
 
             switch (livyState)
             {
-                case "starting":
                 case "error":
                 case "dead":
                 case "success":