+++ sdk/synapse/Azure.Analytics.Synapse.Spark/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release History
 
+## 1.0.0-preview.9 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
 ## 1.0.0-preview.8 (2021-10-05)
 
 ### Features Added